refactor: move animation utilities into Helpers folder
