+++ src/TriQue.App/Forms/AdminForm.cs
@@ -8,9 +8,9 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
-using TriQue;
 using TriQue.Data.Repositories;
 using TriQue.Forms;
+using TriQue.Helpers.Animation;
 using TriQue.Services;
 namespace Trique.Forms
 {
