+++ src/TriQue.App/Forms/AdminSettings.cs
@@ -2,9 +2,9 @@
 using System.Data;
 using System.Drawing;
 using System.Windows.Forms;
-using TriQue;
 using TriQue.Data.Repositories;
 using TriQue.Forms;
+using TriQue.Helpers.Animation;
 using TriQue.Services;
 
 namespace Trique.Forms
