+++ src/TriQue.App/Forms/AdminViewQueue.cs
@@ -1,6 +1,6 @@
-using TriQue;
 using TriQue.Data.Repositories;
 using TriQue.Forms;
+using TriQue.Helpers.Animation;
 using TriQue.Services;
 
 namespace Trique.Forms
