+++ src/TriQue.App/Forms/DriverForm.cs
@@ -2,6 +2,7 @@ using Microsoft.VisualBasic.Logging;
 using TriQue.Data.Repositories;
 using TriQue.DTOs;
 using TriQue.Enums;
+using TriQue.Helpers.Animation;
 using TriQue.Services;
 
 namespace TriQue.Forms
