+++ src/TriQue.App/Helpers/Animation/ModalAnimator.cs
